requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
docker==4.4.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
